embedding_service/requirements.txt
Flask==3.0.0
gunicorn==21.2.0
sentence-transformers==2.5.1
numpy==1.26.0
python-dotenv==1.0.0
